Allow const and unit enum variants in test_case (#16)
* Properly typecheck arguments in test_case.

Previously, the type of the function signature would be ignored by the
function generated by test_case. This meant that the following would
work:

```
fn test(a: u32) {
    let b: u64 = a;
}
```

This is because test_case would generate this function, which would
compile properly without errors:

```
fn test_1() {
    let a = 1;
    let b: u64 = a;
}
```

This is highly surprising, and can lead to some strange,
hard-to-understand behavior in the test cases.

We now specify the type when we assign `a`, e.g. the assignment becomes
`let a: u32 = 1`, ensuring a nice error message is shown:

```
error[E0308]: mismatched types
  --> ntest/tests/integration.rs:19:50
   |
19 |     let b: u64 = a;
   |                  ^ expected `u64`, found `u32`
   |
help: you can convert a `u32` to a `u64`
   |
19 |     let b: u64 = a.into();
   |                   +++++++
```

* Allow const and unit enums/struct in test_case.

This commit allows passing a Path to the test_case macro, paving the way
to pass constants, unit enum variants, and unit structs.